refactor: move healtha nd mock transport to separate files
