+++ pkg/es/client.go
@@ -1,12 +1,9 @@
 package es
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 
 	"github.com/elastic/go-elasticsearch/v7"
-	"github.com/elastic/go-elasticsearch/v7/esapi"
 	"github.com/esctl/esctl/pkg/config"
 )
 
@@ -29,25 +26,4 @@ func New(c *config.Cluster) (*elasticSearchClient, error) {
 
 type elasticSearchClient struct {
 	c *elasticsearch.Client
-}
-
-func (e *elasticSearchClient) GetHealth() (HealthResponse, error) {
-	healthResponse := HealthResponse{}
-	req := esapi.ClusterHealthRequest{}
-	res, err := req.Do(context.Background(), e.c)
-	if err != nil {
-		return healthResponse, err
-	}
-
-	err = json.NewDecoder(res.Body).Decode(&healthResponse)
-	if err != nil {
-		return healthResponse, err
-	}
-
-	return healthResponse, nil
-}
-
-type HealthResponse struct {
-	ClusterName string `json:"cluster_name"`
-	Status      string
 }
